requirements.txt
pandas == 2.2.3
openpyxl == 3.1.5
python-dotenv == 1.1.0
twilio == 9.5.2